@@ -17,30 +17,28 @@ import android.view.View;
 import android.widget.EditText;
 import android.widget.ImageButton;
 import android.widget.ProgressBar;
-import android.widget.TextView;
 import android.widget.Toast;
 
-import com.google.gson.Gson;
-import com.nanorep.convesationui.ConversationViewsProvider;
-import com.nanorep.convesationui.adapter.ConversationInjector;
 import com.nanorep.convesationui.fragments.NRConversationFragment;
+import com.nanorep.convesationui.structure.ChatEventListener;
+import com.nanorep.convesationui.structure.ChatServicesInjector;
 import com.nanorep.convesationui.structure.FriendlyDatestampFormatFactory;
 import com.nanorep.convesationui.structure.OptionActionListener;
 import com.nanorep.convesationui.structure.UiConfigurations;
-import com.nanorep.convesationui.structure.history.HistoryHandler;
-import com.nanorep.convesationui.structure.history.HistoryItem;
-import com.nanorep.convesationui.structure.history.HistoryItemOrigin;
 import com.nanorep.convesationui.structure.history.HistoryListener;
 import com.nanorep.convesationui.structure.history.HistoryProvider;
 import com.nanorep.convesationui.viewholder.base.ChatElementViewHolder;
 import com.nanorep.convesationui.viewholder.controllers.ChatElementController;
+import com.nanorep.convesationui.views.AgentTypingSettingsProvider;
 import com.nanorep.convesationui.views.carousel.CarouselInfoContainer;
 import com.nanorep.convesationui.views.carousel.CarouselItemConfiguration;
 import com.nanorep.convesationui.views.carousel.CarouselItemsContainer;
 import com.nanorep.convesationui.views.carousel.CarouselView;
 import com.nanorep.convesationui.views.carousel.CarouselViewHolder;
-import com.nanorep.convesationui.views.carousel.CarouselViewsProvider;
-import com.nanorep.nanoengine.ConversationListener;
+import com.nanorep.convesationui.views.providers.CarouselViewsProvider;
+import com.nanorep.convesationui.views.providers.ConversationViewsProvider;
+import com.nanorep.convesationui.views.providers.FeedbackUIProvider;
+import com.nanorep.nanoengine.ConversationStateListener;
 import com.nanorep.nanoengine.Entity;
 import com.nanorep.nanoengine.NRConversationEngine;
 import com.nanorep.nanoengine.NRConversationMissingEntities;
@@ -48,8 +46,10 @@ import com.nanorep.nanoengine.NRPrivateInfo;
 import com.nanorep.nanoengine.NanoAccess;
 import com.nanorep.nanoengine.NanoRep;
 import com.nanorep.nanoengine.Property;
+import com.nanorep.nanoengine.chatelement.ChatElement;
+import com.nanorep.nanoengine.chatelement.StorableChatElement;
 import com.nanorep.nanoengine.model.AgentType;
-import com.nanorep.nanoengine.model.CarouselData;
+import com.nanorep.nanoengine.model.CarouselElement;
 import com.nanorep.nanoengine.model.NRAccount;
 import com.nanorep.nanoengine.model.configuration.ConversationSettings;
 import com.nanorep.nanoengine.model.configuration.TimestampStyle;
@@ -58,8 +58,10 @@ import com.nanorep.nanoengine.model.conversation.statement.OnStatementResponse;
 import com.nanorep.nanoengine.model.conversation.statement.StatementRequest;
 import com.nanorep.nanoengine.model.conversation.statement.StatementResponse;
 import com.nanorep.nanoengine.model.conversation.statement.StatementScope;
+import com.nanorep.sdkcore.model.StatementStatus;
 import com.nanorep.sdkcore.types.NRError;
 import com.nanorep.sdkcore.types.ViewHolder;
+import com.nanorep.sdkcore.utils.UtilsKt;
 import com.nanorep.sdkcore.utils.network.ConnectivityReceiver;
 
 import org.jetbrains.annotations.NotNull;
@@ -72,29 +74,35 @@ import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
+import java.util.Random;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.CopyOnWriteArrayList;
 
-import static com.nanorep.nanoengine.model.NRChannel.ChannelType.Chat;
-import static com.nanorep.nanoengine.model.NRChannel.ChannelType.PhoneNumber;
-import static com.nanorep.nanoengine.model.NRChannel.ChannelType.Ticket;
-import static com.nanorep.nanoengine.model.ResponseModelsKt.isConversationError;
+import kotlin.Unit;
+import kotlin.jvm.functions.Function1;
+
+import static com.nanorep.nanoengine.model.Channels.Chat;
+import static com.nanorep.nanoengine.model.Channels.PhoneNumber;
+import static com.nanorep.nanoengine.model.Channels.Ticket;
 import static com.nanorep.sdkcore.model.CommunicationModelsKt.StatusOk;
+import static com.nanorep.sdkcore.model.CommunicationModelsKt.StatusPending;
 import static com.nanorep.sdkcore.utils.UtilsKt.getPx;
 import static com.nanorep.sdkcore.utils.UtilsKt.snack;
 import static com.nanorep.sdkcore.utils.UtilsKt.toast;
+import static java.security.AccessController.getContext;
 
-public class MainActivity extends AppCompatActivity implements ConversationListener, HistoryProvider, ConnectivityReceiver.ConnectivityListener {
+public class MainActivity extends AppCompatActivity implements
+        ConnectivityReceiver.ConnectivityListener,
+        ConversationStateListener, HistoryProvider, ChatEventListener {
 
     private static final String TAG = "MainActivity";
 
-
     /* --- Please fill with valid account values --- */
-    private static final String ACCOUNT_NAME = BuildConfig.ACCOUNT_NAME;
-    private static final String API_KEY = BuildConfig.API_KEY;
-    private static final String KNOWLEDGE_BASE = BuildConfig.KNOWLEDGE_BASE;
-    private static final String SERVER = BuildConfig.SERVER;
+    private static final String ACCOUNT_NAME = "";
+    private static final String API_KEY = "";
+    private static final String KNOWLEDGE_BASE = "";
+    private static final String SERVER = "";
     /*///////////////////////////////////////////////*/
 
 
@@ -104,13 +112,12 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
 
     private ImageButton startButton;
     private ProgressBar progressBar;
-    private TextView historySizeTextView;
 
     private NRAccount account;
     private String conversationId;
     private NanoAccess nanoAccess;
 
-    private ConcurrentHashMap<String, List<HistoryItem>> chatHistory = new ConcurrentHashMap<>();
+    private ConcurrentHashMap<String, List<HistoryElement>> chatHistory = new ConcurrentHashMap<>();
     final Object historySync = new Object(); // in use to block multi access to history from different actions.
 
     private Map<String, OpenConversation> openConversations = new HashMap<>();
@@ -119,7 +126,6 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
     private ConnectivityReceiver connectivityReceiver = new ConnectivityReceiver();
     private ConcurrentLinkedQueue<StatementRequest> failedStatements = new ConcurrentLinkedQueue<>();
     private SoftReference<NRConversationFragment> conversationFragment;
-    private HistoryHandler historyHandler;
 
     /**
      * in use when previously failed statements are posted to indicate if the connection is done again
@@ -141,10 +147,11 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
     /**
      * will be used when posting failed statements
      */
-    private OnStatementResponse onStatementResponse = new OnStatementResponse() {
+    private OnStatementResponse onStatementResponse = new OnStatementResponse.Defaults() {
+
         @Override
         public void onResponse(@NotNull StatementResponse response) {
-            if(conversationFragment.get() != null){
+            if (conversationFragment.get() != null) {
                 conversationFragment.get().onResponse(response);
 
             } else {
@@ -154,18 +161,17 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
 
                 // update request history item with the new status.
                 StatementRequest statementRequest = response.getStatementRequest();
-                if(statementRequest != null) {
-                    historyHandler.updateHistory(statementRequest.getTimestamp(), statementRequest.getTimestamp(), StatusOk);
+                if (statementRequest != null) {
+                    update(statementRequest.getTimestamp(), statementRequest.getTimestamp(), StatusOk);
                 }
                 // add response history item to the history storage
-                historyHandler.addToHistory(new HistoryItem(new Gson().toJson(response),
-                        response.getTimestamp(), HistoryItemOrigin.INCOMING, StatusOk));
+                store(StorableChatElement.StorableFactory.create(response));
             }
         }
 
         @Override
         public void onError(@NotNull NRError error) {
-            Log.e("MainFragment","failed to re-post request: "+error.getErrorCode()+", "+error.getReason() );
+            Log.e("MainFragment", "failed to re-post request: " + error.getErrorCode() + ", " + error.getReason());
             MainActivity.this.onError(error);
         }
     };
@@ -176,27 +182,18 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
         setContentView(R.layout.activity_main);
 
         startButton = findViewById(R.id.startButton);
-        //historySizeTextView = findViewById(R.id.historySizeTextView);
         progressBar = findViewById(R.id.progressBar);
-        updateHistorySizeIndication();
 
-        startButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                setLoadingDisplay(true);
-                initNanorep();
-            }
+        startButton.setOnClickListener(view -> {
+            setLoadingDisplay(true);
+            initNanorep();
         });
 
-        getSupportFragmentManager().addOnBackStackChangedListener(new FragmentManager.OnBackStackChangedListener() {
-            @Override
-            public void onBackStackChanged() {
-                if (getSupportFragmentManager().getBackStackEntryCount() == 0) {
-                    setLoadingDisplay(false);
-                    updateHistorySizeIndication();
-                    if(conversationFragment != null) {
-                        conversationFragment.clear();
-                    }
+        getSupportFragmentManager().addOnBackStackChangedListener(() -> {
+            if (getSupportFragmentManager().getBackStackEntryCount() == 0) {
+                setLoadingDisplay(false);
+                if(conversationFragment != null) {
+                    conversationFragment.clear();
                 }
             }
         });
@@ -212,10 +209,6 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
         super.onStart();
 
         connectivityReceiver.register(this, this);
-    }
-
-    private void updateHistorySizeIndication() {
-        //historySizeTextView.setText(getString(R.string.history_size_format, getHistorySize()));
     }
 
     @Override
@@ -246,14 +239,15 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
         Map<String, String> conversationContext = new HashMap<>();
         account = getAccount();
         account.setContext(conversationContext);
-        account.setEntities(new String[]{"SUBSCRIBERS"});
+        account.setEntities(new String[]{"SUBSCRIBER"});
 
-        ConversationSettings settings = new ConversationSettings().disableFeedback()
-                .speechEnable(true).enableMultiRequestsOnLiveAgent(true)
-                .enableOfflineMultiRequests(true)
-                .timestampConfig(true, new TimestampStyle("hh:mm a",
-                        getPx(10), Color.parseColor("#a8a8a8"), null) ).
-                        datestamp(true, new FriendlyDatestampFormatFactory(this));
+        ConversationSettings settings = new ConversationSettings()
+                .speechEnable(true).enableMultiRequestsOnLiveAgent(true).setReadMoreThreshold(320)
+                .timestampConfig(true, new TimestampStyle("hh:mm aaa",
+                        getPx(11), Color.parseColor("#33aa33"), null))
+                .enableOfflineMultiRequests(true) // defaults to true
+                .enableLiveAgent(true)
+                .datestamp(true, new FriendlyDatestampFormatFactory(this));
 
         //-> to configure the default text configuration to all bubbles text add something like the following:
         //settings.textStyleConfig(new StyleConfig(getPx(23), Color.BLUE, null));
@@ -270,7 +264,7 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
             apiKey = ((EditText) findViewById(R.id.api_key_edit_text)).getText().toString();
             kb = ((EditText) findViewById(R.id.knowledgebase_edit_text)).getText().toString();
             server = ((EditText) findViewById(R.id.server_edit_text)).getText().toString();
-        }catch (Exception ex){}
+        }catch (Exception ignored){}
 
         return new NRAccount(accountName.isEmpty() ? ACCOUNT_NAME : accountName,
                 apiKey.isEmpty() ? API_KEY : apiKey,
@@ -284,7 +278,10 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
 
         if (openConversations.containsKey(account)) {
             OpenConversation openConversation = openConversations.get(account);
-            conversation = new Conversation(openConversation.conversationId, openConversation.lastAgent);
+
+            if (openConversation != null) {
+                conversation = new Conversation(openConversation.conversationId, openConversation.lastAgent);
+            }
         }
 
         if (conversation == null) {
@@ -303,25 +300,21 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
             return;
         }
 
-        this.conversationFragment = new SoftReference<>(NRConversationFragment.newInstance(new ConversationInjector() {
-            @Override
-            public ConversationListener getListener() {
-                return MainActivity.this;
-            }
+        this.conversationFragment = new SoftReference<>(
+                NRConversationFragment.newInstance(new ChatServicesInjector() {
 
-            @Override
-            public ConversationViewsProvider getUiProvider() {
-                return new MyConversationUiProvider();
-            }
+                    @Override
+                    public HistoryProvider getHistoryProvider() {
+                        return MainActivity.this;
+                    }
 
+                    @Override
+                    public ConversationViewsProvider getUiProvider() {
+                        return new MyConversationUiProvider();
+                    }
 
-            @Override
-            public HistoryHandler getHistoryHandler() {
-                historyHandler = new HistoryHandler(MainActivity.this, true);
-                return historyHandler;
-//                 return new ConversationInjector.DefaultsInjector().getHistoryHandler();// - if history is not needed
-            }
-        }));
+                }, MainActivity.this)
+        );
 
         fragmentManager
                 .beginTransaction()
@@ -405,11 +398,11 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
     }
 
     @Override
-    public void onConversationTermination(Conversation conversation) {
+    public void onConversationTermination(@NotNull Conversation conversation) {
 
-        OpenConversation openConversation;
+       OpenConversation openConversation;
         String accountId = this.account.getAccount();
-        if(openConversations.containsKey(accountId)){
+        if (openConversations.containsKey(accountId)) {
             openConversation = openConversations.get(accountId);
             openConversation.conversationId = conversation.getId();
             openConversation.lastAgent = conversation.getAgentType();
@@ -419,18 +412,26 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
         }
     }
 
+    /**
+     * Missing entity contains extra data - "Properties" that are being provided by the app (which can bring the data from external resources).
+     * For example: If the provider contains Missing entity called "SUBSCRIBER", the app is being notified here and returns the extra data.
+     *
+     * @param response - The response from nanorep that contains the Missing Entities
+     * @param missingEntitiesHandler - Would be informed when the missing entities are ready to be injected to the provider
+     */
     @Override
-    public void onMissingEntities(StatementResponse response, NRConversationEngine.MissingEntitiesHandler handler) {
+    public void onMissingEntities(@NotNull StatementResponse response, @NotNull NRConversationEngine.MissingEntitiesHandler missingEntitiesHandler) {
         NRConversationMissingEntities missingEntities = new NRConversationMissingEntities(response.getStatement());
         for (String missingEntity : response.getMissingEntities()) {
-            if (missingEntity.equals("SUBSCRIBERS")) {
+            if (missingEntity.equals("SUBSCRIBER")) {
                 missingEntities.addEntity(createEntity(missingEntity));
             }
         }
 
-        handler.onMissingEntitiesReady(missingEntities);
+        missingEntitiesHandler.onMissingEntitiesReady(missingEntities);
     }
 
+    // For this example, the app returns random data:
     private Entity createEntity(String entityName) {
         Entity entity = new Entity(Entity.PERSISTENT, Entity.NUMBER, "3", entityName, "1");
         for (int i = 0; i < 3; i++) {
@@ -442,8 +443,9 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
         return entity;
     }
 
+
     @Override
-    public void handlePersonalInformation(String id, NRPrivateInfo privateInfo) {
+    public void handlePersonalInformation(@NotNull String id, @NotNull NRPrivateInfo privateInfo) {
         switch (id) {
             case "balance":
                 String balance = String.format("%10.2f$", Math.random() * 10000);
@@ -457,15 +459,14 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
     @Override
     public void onInitializeChatHandover() {
         handoverReplyCount = 0;
-        if(!connectionOk){
+        if (!connectionOk) {
             nanoAccess.endHandover();
             // response scope should be provided indicate the chat agent status while statement were injected.
             // (see StatementScope for available scopes)
-            conversationFragment.get().injectResponse("Failed to initiate live chat, please check your internet connection",
-                    StatementScope.SystemScope());
+            conversationFragment.get().injectSystem("Failed to initiate live chat, " +
+                    "please check your internet connection");
         } else {
-            conversationFragment.get().injectResponse("Hey, my name is Bob, how can I help?",
-                    StatementScope.HandoverScope());
+            passDelayedResponse("Hey, my name is Bob, how can I help?", 1);
         }
     }
 
@@ -496,16 +497,25 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
         }
     }
 
-    private boolean isEndHandover(String inputText) {
-        return inputText != null && inputText.equalsIgnoreCase(END_HANDOVER_SESSION);
+    private void passDelayedResponse(final String responseText, final int count) {
+        new Thread(new Runnable() {
+            @Override
+            public void run() {
+                int random = count != -1 ? count-1  : new Random().nextInt(3);
+                for (int i = 0; i <= random; i++) {
+                    passHandoverResponse(responseText + (i == 0 ? "" : "_" + i), 1000);
+                }
+            }
+        }).start();
     }
 
     private void passHandoverResponse(@NonNull StatementRequest statementRequest, boolean endHandover) {
 
-        String responseText = "";
+        // pass live agent response to the chat SDK
+        nanoAccess.updateRequestStatus(statementRequest, StatusOk);
 
         if (endHandover) {
-            responseText = "Bye - Handover complete";
+            passHandoverResponse("Bye - Handover complete", 0);
 
         } else {
             //!- for long text test:
@@ -513,15 +523,48 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
                     "The good news is that humanity finally made it into interstellar space. The bad news is that planets fit to live on are scarce - and alien races willing to fight us for them are common. So, we fight, to defend Earth and to stake our own claim to planetary real estate. Far from Earth, the war has been going on for decades: brutal, bloody, unyielding.\n" +
                     "\n" +
                     "Earth itself is a backwater. The bulk of humanity's resources are in the hands of the Colonial Defense Force. Everybody knows that when you reach retirement age, you can join the CDF. They don't want young people; they want people who carry the knowledge and skills of decades of living. You'll be taken off Earth and never allowed to return. You'll serve two years at the front. And if you survive, you'll be given a generous homestead stake of your own, on one of our hard-won colony planets. ";
-            String postText = handoverReplyCount == 1 ? longText : " ";
-            responseText = postText +"Response (" + String.valueOf(++handoverReplyCount) + ")";
+            String postText = handoverReplyCount == 1 ? longText : "";
+            final String responseText = postText + "Response (" + String.valueOf(++handoverReplyCount) + ")";
+
+            passDelayedResponse(responseText, postText.isEmpty() ? -1 : 1);
+        }
+    }
+
+    private void passHandoverResponse(final String response, int delay) {
+
+        // Present typing indication
+        UtilsKt.runInMain(nanoAccess, new Function1<NanoAccess, Unit>() {
+            @Override
+            public Unit invoke(NanoAccess nanoAccess) {
+                nanoAccess.updateTypingState(true);
+                return null;
+            }
+        });
+
+        if (delay != 0) {
+            try {
+                Thread.sleep(delay);
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            }
         }
 
-        // pass live agent response to the chat SDK
-        if(statementRequest.getCallback() != null){
-            statementRequest.getCallback().onResponse(new StatementResponse(responseText, statementRequest));
-        }
+        // Hide typing indication
+        UtilsKt.runInMain(nanoAccess, new Function1<NanoAccess, Unit>() {
+            @Override
+            public Unit invoke(NanoAccess nanoAccess) {
+                nanoAccess.updateTypingState(false);
 
+                if (conversationFragment.get() != null) {
+                    conversationFragment.get().injectResponse(response, StatementScope.HandoverScope());
+                }
+                return null;
+            }
+        });
+    }
+
+    private boolean isEndHandover(String inputText) {
+        return inputText != null && inputText.equalsIgnoreCase(END_HANDOVER_SESSION);
     }
 
 //</editor-fold>
@@ -529,11 +572,12 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
 
 //<editor-fold desc=">>>>>>>>> Error handling and ErrorListener Impl  <<<<<<<<<">
 
+    @SuppressLint("ResourceType")
     @Override
-    public void onError(NRError error) {
+    public void onError(@NotNull NRError error) {
         progressBar.setVisibility(View.INVISIBLE);
 
-        switch (error.getErrorCode()){
+        switch (error.getErrorCode()) {
 
             case NRError.ConversationCreationError:
 
@@ -541,24 +585,27 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
 
                 pendingConversationCreation = true;
 
-                if(NRError.ConnectionException.equals(error.getReason())) {
-                    if(conversationFragment == null || conversationFragment.get() == null) {
+                if (NRError.ConnectionException.equals(error.getReason())) {
+                    if (conversationFragment == null || conversationFragment.get() == null) {
                         openConversationFragment();
-                        /*!! after opening the fragment, we need to wait for
-                          !! onChatLoaded event before activating anything on it.*/
+
+                        //!! request injection is not allowed right after fragment committing to backstack
+                        // that will result with StatementError. wait to the onChatLoaded event
+                        // (see onChatLoaded)
                     }
                 }
 
-                if(startButton!=null){
-                    startButton.setEnabled(true);
+                View accountBtn = getConversationStartBtn();
+                if (accountBtn != null) {
+                    accountBtn.setEnabled(true);
                 }
                 break;
 
             case NRError.StatementError:
 
-                if(conversationFragment == null || conversationFragment.get() == null) return;
+                if (conversationFragment == null || conversationFragment.get() == null) return;
 
-                if(isConversationError(error)){
+                if (error.isConversationError()) {
                     notifyConversationError(error);
                     pendingConversationCreation = true;
 
@@ -566,13 +613,11 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
                     notifyStatementError(error);
                 }
                 StatementRequest statementRequest = (StatementRequest) error.getData();
-                if(statementRequest != null) {
-                /*
-                Log.d("MainFragment", "removing failed request "+statementRequest.getStatement() +" from History");
-                historyHandler.removeFromHistory(statementRequest.getTimestamp());
-                 */
+                if (statementRequest != null) {
+                    //Log.d("MainFragment", "removing failed request " + statementRequest.getStatement() + " from History");
+                    //removeFromHistory(statementRequest.getTimestamp());
 
-                    Log.d("MainFragment", "adding "+statementRequest.getStatement() +" to app pending");
+                    Log.d("MainFragment", "adding " + statementRequest.getStatement() + " to app pending");
                     failedStatements.add(statementRequest);
                 }
                 break;
@@ -583,9 +628,9 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
                  if needed the error.getErrorCode() and sometimes the error.getReason() can provide
                  the details regarding the error
                  */
-                if(!isFinishing()){
+                if (getContext() != null) {
                     Log.e("App-ERROR", error.toString());
-                    toast(this, error.toString(), 2500 );
+                    toast(this, error.toString(), 2500);
                 }
         }
     }
@@ -611,8 +656,11 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
         }
     }
 
-//</editor-fold>
+    View getConversationStartBtn() {
+        return startButton;
+    }
 
+//</editor-fold>
 
 //<editor-fold desc=">>>>>>>>>  ChatEventListener impl  <<<<<<<<<">
 
@@ -649,9 +697,7 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
             Intent intent = new Intent(Intent.ACTION_DIAL);
             intent.setData(Uri.parse("tel:" + phoneNumber));
             startActivity(intent);
-        } catch (ActivityNotFoundException e) {
-
-        }
+        } catch (ActivityNotFoundException ignored) {}
     }
 
 //</editor-fold>
@@ -659,52 +705,68 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
 
 //<editor-fold desc=">>>>>>>>>  History handling and HistoryProvider Impl  <<<<<<<<<">
 
+    private int getHistorySize() {
+        String accountId = this.account != null ? this.account.getAccount() : null;
+        return accountId != null && chatHistory.containsKey(accountId) ? chatHistory.get(accountId).size() : 0;
+    }
+
     @Override
-    public void fetchHistory(final int from, final boolean older, final HistoryListener listener) {
-        new Thread(new Runnable() {
-            @Override
-            public void run() {
-                final List<HistoryItem> history;
+    public void onFileUploadClicked() {
+        Toast.makeText(this, "File upload clicked", Toast.LENGTH_SHORT).show();
+    }
 
-                synchronized (historySync) {
-                    history = Collections.unmodifiableList(getHistoryForAccount(account.getAccount(), from, older));
-                }
+    @Override
+    public void fetch(int from, int direction, @Nullable HistoryListener listener) {
 
-                if (history.size() > 0) {
-                    try {
-                        Thread.sleep(1000); // simulate async history fetching
-                    } catch (InterruptedException e) {
-                        e.printStackTrace();
-                    }
-                }
+        new Thread(() -> {
+            final List<? extends StorableChatElement> history;
 
-                if(!isFinishing()) {
-                    runOnUiThread(new Runnable() {
-                        @Override
-                        public void run() {
-                            listener.onReady(from, older, history);
-                        }
-                    });
+            synchronized (historySync) {
+                history = Collections.unmodifiableList(getHistoryForAccount(account.getAccount(), from, direction));
+            }
+
+            if (history.size() > 0) {
+                try {
+                    Thread.sleep(800); // simulate async history fetching
+                } catch (InterruptedException e) {
+                    e.printStackTrace();
                 }
             }
+            runOnUiThread(() -> {
+                Log.d("History", "passing history list to listener, from = " + from + ", size = " + history.size());
+                if (listener != null) {
+                    listener.onReady(from, direction, history);
+                }
+            });
         }).start();
     }
 
-    private List<HistoryItem> getHistoryForAccount(String account, int fromIdx, boolean older) {
-        if(!chatHistory.containsKey(account)) return new ArrayList<>();
+    private List<HistoryElement> getHistoryForAccount(String account, int fromIdx, int direction) {
+
+        List<HistoryElement> accountChatHistory = chatHistory.get(account);
+
+        if (accountChatHistory == null)
+            return new ArrayList<>();
+
+        boolean fetchOlder = direction == Older;
 
         // to prevent Concurrent exception
-        CopyOnWriteArrayList<HistoryItem> accountHistory = new CopyOnWriteArrayList<>(chatHistory.get(account));
+        CopyOnWriteArrayList<HistoryElement> accountHistory = new CopyOnWriteArrayList<>(accountChatHistory);
 
-        if(fromIdx == -1) {
-            fromIdx = accountHistory.size() - 1;
-        } else {
-            fromIdx = accountHistory.size() - fromIdx;
+        int historySize = accountHistory.size();
+
+        if (fromIdx == -1) {
+            fromIdx = fetchOlder ? historySize - 1 : 0;
+        } else if (fetchOlder) {
+            fromIdx = historySize - fromIdx;
         }
 
-        int toIndex = Math.max(0, fromIdx-HistoryPageSize);
+        int toIndex = fetchOlder ? Math.max(0, fromIdx - HistoryPageSize) :
+                Math.min(fromIdx + HistoryPageSize, historySize - 1);
 
         try {
+            Log.d("History", "fetching history items (" + historySize + ") from " + toIndex + " to " + fromIdx);
+
             return accountHistory.subList(toIndex, fromIdx);
 
         } catch (Exception ex) {
@@ -712,129 +774,175 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
         }
     }
 
-
-    @Override
-    public void storeToHistory(HistoryItem item) {
-        //if(item == null || item.getStatus() != StatusOk) return;
-
-        ArrayList<HistoryItem> convHistory = getAccountHistory();
-        convHistory.add(item);
-    }
-
-    @Override
-    public void removeFromHistory(long timestampId) {
-        ArrayList<HistoryItem> convHistory = getAccountHistory();
-
-        Iterator<HistoryItem> iterator = convHistory.listIterator();
-        while (iterator.hasNext()) {
-            HistoryItem item = iterator.next();
-            if (item.getTimestamp() == timestampId) {
-                iterator.remove();
-                break;
-            }
-        }
-    }
-
-    @Override
-    public void updateHistory(long timestampId, long newTimestamp, int status) {
-        ArrayList<HistoryItem> convHistory = getAccountHistory();
-
-        Iterator<HistoryItem> iterator = convHistory.listIterator();
-        while (iterator.hasNext()) {
-            HistoryItem item = iterator.next();
-            if(item.getTimestamp() == timestampId){
-                item.setTimestamp(newTimestamp);
-                item.setStatus(status);
-                break;
-            }
-        }
-    }
-
     @NonNull
-    private ArrayList<HistoryItem> getAccountHistory() {
-        ArrayList<HistoryItem> convHistory;
+    private ArrayList<HistoryElement> getAccountHistory() {
+        ArrayList<HistoryElement> convHistory;
         String account = this.account.getAccount();
-        if(chatHistory.containsKey(account)) {
-            convHistory = (ArrayList<HistoryItem>) chatHistory.get(account);
+
+        if (chatHistory != null && chatHistory.containsKey(account)) {
+            convHistory = (ArrayList<HistoryElement>) chatHistory.get(account);
         } else {
-            convHistory = new ArrayList<HistoryItem>();
-            chatHistory.put(account,  convHistory);
+            convHistory = new ArrayList<>();
+            chatHistory.put(account, convHistory);
         }
         return convHistory;
     }
 
-    private int getHistorySize() {
-        String accountId = this.account != null ? this.account.getAccount() : null;
-        return accountId != null && chatHistory.containsKey(accountId) ? chatHistory.get(accountId).size() : 0;
+    @Override
+    public void store(@NotNull StorableChatElement item) {
+        //if(item == null || item.getStatus() != StatusOk) return;
+
+        synchronized (historySync) {
+            ArrayList<HistoryElement> convHistory = getAccountHistory();
+            convHistory.add(new HistoryElement(item));
+        }
+    }
+
+    @Override
+    public void remove(long timestampId) {
+        synchronized (historySync) {
+            ArrayList<HistoryElement> convHistory = getAccountHistory();
+
+            Iterator<HistoryElement> iterator = convHistory.listIterator();
+            while (iterator.hasNext()) {
+                HistoryElement item = iterator.next();
+                if (item.getTimestamp() == timestampId) {
+                    iterator.remove();
+                    break;
+                }
+            }
+        }
+    }
+
+    @Override
+    public void update(long timestampId, long newTimestamp, int status) {
+
+        synchronized (historySync) {
+            ArrayList<HistoryElement> convHistory = getAccountHistory();
+
+            for (HistoryElement item : convHistory) {
+                if (item.getTimestamp() == timestampId) {
+                    item.timestamp = newTimestamp;
+                    item.status = status;
+                    break;
+                }
+            }
+        }
+    }
+
+    @Override
+    public void update(long timestampId, @NotNull StorableChatElement item) {
+        synchronized (historySync) {
+            ArrayList<HistoryElement> convHistory = getAccountHistory();
+            for (int i = convHistory.size() - 1; i >= 0; i--) {
+                if (convHistory.get(i).getTimestamp() == timestampId) {
+                    convHistory.set(i, new HistoryElement(item));
+                    break;
+                }
+            }
+                /* -> Another option:
+                for (HistoryElement historyElement: convHistory) {
+                    if (historyElement.getTimestamp() == timestampId) {
+                        historyElement.scope = item.getScope();
+                        historyElement.status = item.getStatus();
+                        historyElement.key = item.getStorageKey();
+                        historyElement.type = item.getType();
+                        break;
+                    }
+                }*/
+        }
     }
 
 //</editor-fold>
 
+//<editor-fold desc=">>>>>>>>>  Carousel overriding implementations:  <<<<<<<<<">
 
-//<editor-fold desc=">>>>>>>>>  Carousel overriding  <<<<<<<<<">
-
-    class MyCarouselUiProvider extends CarouselViewsProvider {
-        @Nullable
-        @Override
-        public ViewHolder getCarouselInfoHolder(@NotNull CarouselData carouselData, @Nullable CarouselInfoContainer carouselInfoContainer) {
-            DemoCarouselInfoViewHolder demoCarouselInfoViewHolder = new DemoCarouselInfoViewHolder(carouselInfoContainer);
-            demoCarouselInfoViewHolder.update(carouselData);
-            return demoCarouselInfoViewHolder;
-        }
-
-        @NotNull
-        @Override
-        public CarouselInfoContainer getCarouselInfoView(@NotNull Context context) {
-            CarouselInfoContainer carouselInfoContainer = super.getCarouselInfoView(context);
-            carouselInfoContainer.setTextBackground(context.getResources().getDrawable(R.drawable.bg_inbox_normal));
-            carouselInfoContainer.setMargins( 0, 0, getPx(30),0);
-            carouselInfoContainer.setIconAlignment(UiConfigurations.Alignment.AlignTop);
-
-            // in case carousel timestamp needs to be changed:
-            // carouselInfoContainer.setTimestampStyle(new TimestampStyle("hh:mm", getPx(12), Color.parseColor("#a8a8a8"), null));
-
-            // carouselInfoContainer.setIconTextGap(getPx(8)); // change the margin between the bot icon and the bubble text
-
-            return carouselInfoContainer;
-        }
-
-        @NotNull
-        @Override
-        public CarouselViewHolder injectCarousel(@NotNull Context context, @Nullable CarouselData carouselData, @Nullable OptionActionListener optionListener) {
-            CarouselView carouselView = (CarouselView) super.injectCustomCarousel(context, carouselData, null);
-            ((CarouselView.LayoutParams) carouselView.getLayoutParams()).setMargins(0, getPx(12), 0, 0);
-            return carouselView;
-        }
-
-        @NotNull
-        @Override
-        public CarouselItemsContainer getCarouselItemsContainer(@NotNull Context context, @Nullable OptionActionListener optionListener) {
-
-            CarouselItemsContainer itemsContainer = super.getCarouselItemsContainer(context, optionListener);
-            itemsContainer.setOptionsHorizontalAlignment(Gravity.END);
-            itemsContainer.setOptionsVerticalAlignment(Gravity.BOTTOM);
-            itemsContainer.setOptionsBackground(new ColorDrawable(Color.TRANSPARENT));
-            itemsContainer.setOptionsTextStyle(Color.parseColor("#5EC4B6"), null);
-            itemsContainer.setOptionsTextSize(16);
-
-            itemsContainer.setItemBackground(new ColorDrawable(Color.parseColor("#F5F4F4")));
-            itemsContainer.setCardStyle(getPx(10), getPx(4) );// set the carousel items to have a "card" look. [roundCornersRadius, elevation]
-            itemsContainer.setInfoTextAlignment(CarouselItemConfiguration.ItemInfoAlignment.AlignBelowThumb);
-            itemsContainer.setInfoTextStyle(Color.GRAY, null);
-            itemsContainer.setInfoTextSize(16);
-            itemsContainer.setInfoTextBackground(new ColorDrawable(Color.parseColor("#00000000")));
-            itemsContainer.setInfoSubTitleMinLines(1);
-            return itemsContainer;
-        }
+class MyCarouselViewsProvider extends CarouselViewsProvider {
+    @Nullable
+    @Override
+    public ViewHolder getCarouselInfoHolder(@NotNull CarouselElement carouselData, @Nullable CarouselInfoContainer carouselInfoContainer) {
+        DemoCarouselInfoViewHolder demoCarouselInfoViewHolder = new DemoCarouselInfoViewHolder(carouselInfoContainer);
+        demoCarouselInfoViewHolder.update(carouselData);
+        return demoCarouselInfoViewHolder;
     }
+
+    @NotNull
+    @Override
+    public CarouselInfoContainer getCarouselInfoView(@NotNull Context context) {
+        CarouselInfoContainer carouselInfoContainer = super.getCarouselInfoView(context);
+        carouselInfoContainer.setTextBackground(context.getResources().getDrawable(R.drawable.bg_inbox_normal));
+        carouselInfoContainer.setMargins(0, 0, getPx(30), 0);
+        carouselInfoContainer.setIconAlignment(UiConfigurations.Alignment.AlignTop);
+        carouselInfoContainer.setIconTextGap(getPx(4)); // change the margin between the bot icon and the bubble text
+        // change the text style of the carousel info section (in bubble)
+        // carouselInfoContainer.setTextStyle(new StyleConfig(getPx(20), Color.BLUE, getTypeface(getContext(), "great_vibes.otf")));
+        // change the carousel timestamp display (if not set, will use the stype provided by ConversationSettings or default if not available)
+        // carouselInfoContainer.setTimestampStyle(new TimestampStyle("hh:mm", getPx(11), Color.parseColor("#a8a8a8"), null));
+
+        return carouselInfoContainer;
+    }
+
+    @NotNull
+    @Override
+    public CarouselViewHolder injectCarousel(@NotNull Context context, @NotNull ChatElementController controller, @Nullable CarouselElement carouselData, @Nullable OptionActionListener optionListener) {
+        CarouselView carouselView = (CarouselView) super.injectCustomCarousel(context, controller, carouselData, null);
+        ((CarouselView.LayoutParams) carouselView.getLayoutParams()).setMargins(0, getPx(16), 0, 0);
+        return carouselView;
+    }
+
+    /**
+     * example to override carousel items list properties.
+     *
+     * @param context
+     * @param optionListener
+     * @return
+     */
+    @NotNull
+    @Override
+    public CarouselItemsContainer getCarouselItemsContainer(@NotNull Context context, @Nullable OptionActionListener optionListener) {
+        CarouselItemsContainer itemsContainer = super.getCarouselItemsContainer(context, optionListener);
+
+        itemsContainer.setOptionsHorizontalAlignment(Gravity.END);
+        itemsContainer.setOptionsVerticalAlignment(Gravity.BOTTOM);
+
+        itemsContainer.setOptionsBackground(/*new ColorDrawable(Color.TRANSPARENT));*/ getResources().getDrawable(R.drawable.c_item_option_back));
+        itemsContainer.setOptionsTextStyle(Color.parseColor("#5EC4B6"), null);
+        itemsContainer.setOptionsTextSize(16);
+
+        itemsContainer.setItemBackground(getResources().getDrawable(R.drawable.bkg_bots));//new ColorDrawable(Color.parseColor("#f1f8ff"))/*getResources().getDrawable(R.drawable.c_item_back)*/);
+        itemsContainer.setCardStyle(getPx(10), getPx(5));// set the carousel items to have a "card" look. [roundCornersRadius, elevation]
+        // itemsContainer.setOptionsTextStyle(Color.parseColor("#00AAFF"), "great_vibes.otf");
+        itemsContainer.setInfoTextAlignment(CarouselItemConfiguration.ItemInfoAlignment.AlignBelowThumb);
+        itemsContainer.setInfoTextStyle(Color.DKGRAY, null);
+        itemsContainer.setInfoTextSize(16);
+        itemsContainer.setInfoTextBackground(new ColorDrawable(Color.parseColor("#00000000")));
+
+        // if the text not occupies the minimum lines, empty lines will be displayed
+        itemsContainer.setInfoTitleMinLines(1); // configure the MINIMUM lines the title will be displayed on
+        itemsContainer.setInfoSubTitleMinLines(2);// configure the MINIMUM lines the title will be displayed on
+        return itemsContainer;
+    }
+}
+
+//</editor-fold>
 
 ///////////////////////////////
 
 
     class MyConversationUiProvider extends ConversationViewsProvider {
 
+        DynamicBubbleBind dynamicBubbleBind = new DynamicBubbleBind();
+
         MyConversationUiProvider() {
-            super(new MyCarouselUiProvider());
+            super(new MyCarouselViewsProvider(), new FeedbackUIProvider());
+                /* !! uncomment this to feedback view override
+                    {
+                @NotNull
+                @Override
+                public FeedbackUIAdapter getFeedbackUIAdapter(@NotNull Context context, int feedbackDisplayType) {
+                    return new FeedbackViewDummy(context);
+                }
+            }*/
         }
 
         /**
@@ -874,12 +982,12 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
 
         @Override
         public ChatElementViewHolder getRemoteBubbleViewHolder(View view, ChatElementController controller) {
-            return new DemoRemoteViewHolder(view, controller);
+            return new DemoRemoteViewHolder(view, controller, dynamicBubbleBind);
         }
 
         @Override
         public ChatElementViewHolder getLocalBubbleViewHolder(View view, ChatElementController controller) {
-            return new DemoLocalViewHolder(view, controller);
+            return new DemoLocalViewHolder(view, controller, dynamicBubbleBind);
         }
 
         @Override
@@ -891,6 +999,17 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
         public int getUserInputLayout() {
             return R.layout.user_input_view_holder_custom;
         }
+
+        @Override
+        public AgentTypingSettingsProvider getAgentTypingSettingsProvider() {
+            AgentTypingSettingsProvider agentTypingSettingsProvider = new AgentTypingSettingsProvider();
+            agentTypingSettingsProvider.setMargins(0, 0, 0, 120);
+            agentTypingSettingsProvider.setGravity(Gravity.BOTTOM);
+            agentTypingSettingsProvider.setCompoundText("please wait");
+            agentTypingSettingsProvider.setCompoundDrawable(R.drawable.ic_more_horiz_black_24dp);
+            return agentTypingSettingsProvider;
+        }
+
     }
 
 //////////////////////////////////////
@@ -907,4 +1026,73 @@ public class MainActivity extends AppCompatActivity implements ConversationListe
             this.lastAgent = lastAgent;
         }
     }
+
+/////////////////////////////////////
+
+    /**
+     * {@link StorableChatElement} implementing class
+     * sample class for app usage
+     */
+    static class HistoryElement implements StorableChatElement {
+        byte[] key;
+        long timestamp = 0;
+        StatementScope scope;
+        @ChatElement.Companion.ChatElementType
+        int type;
+        @StatementStatus
+        int status = StatusPending;
+
+        protected HistoryElement(int type, long timestamp) {
+            this.type = type;
+            this.timestamp = timestamp;
+        }
+
+        HistoryElement(StorableChatElement storable) {
+            key = storable.getStorageKey();
+            type = storable.getType();
+            timestamp = storable.getTimestamp();
+            status = storable.getStatus();
+            scope = storable.getScope();
+        }
+
+        @NotNull
+        @Override
+        public byte[] getStorageKey() {
+            return key;
+        }
+
+        @NotNull
+        @Override
+        public String getStorableContent() {
+            return new String(key);
+        }
+
+        @Override
+        public int getType() {
+            return type;
+        }
+
+        @Override
+        public long getTimestamp() {
+            return timestamp;
+        }
+
+        @Override
+        public int getStatus() {
+            return status;
+        }
+
+        @Override
+        public StatementScope getScope() {
+            return scope;
+        }
+
+        @Override
+        public boolean isStorageReady() {
+            return true;
+        }
+    }
+
+/////////////////////////////////////
+
 }
