@@ -3,7 +3,7 @@ package com.conversation.demo;
 import com.nanorep.convesationui.views.carousel.CarouselInfoContainer;
 import com.nanorep.convesationui.views.carousel.CarouselInfoHolder;
 import com.nanorep.nanoengine.model.AgentType;
-import com.nanorep.nanoengine.model.CarouselData;
+import com.nanorep.nanoengine.model.CarouselElement;
 
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -21,8 +21,8 @@ public class DemoCarouselInfoViewHolder extends CarouselInfoHolder {
 
         super.update(data);
 
-        if (data instanceof CarouselData) {
-            CarouselData carouselData = (CarouselData) data;
+        if (data instanceof CarouselElement) {
+            CarouselElement carouselData = (CarouselElement) data;
 
             int iconId = carouselData.getAgentType() == AgentType.Live ? R.drawable.mr_livechat : R.drawable.mr_chatbot;
             if (getView() != null && getView().getContext() != null) {
